+++ lib/Core/ExecutionState.cpp
@@ -162,9 +162,9 @@ void ExecutionState::deallocate(const MemoryObject *mo) {
 
   auto address = reinterpret_cast<void *>(mo->address);
   if (mo->isLocal) {
-    stackAllocator.free(address, std::max(mo->size, mo->alignment));
+    stackAllocator.free(address, std::max(mo->size, (size_t) mo->alignment));
   } else {
-    heapAllocator.free(address, std::max(mo->size, mo->alignment));
+    heapAllocator.free(address, std::max(mo->size, (size_t) mo->alignment));
   }
 }
